Module lightbeam::microwasm

Modules

sint

Structs

BrTable

BrTargetDrop

Ieee32

Ieee64

MemoryImmediate

MicrowasmConv

OpSig

SignfulInt

Enums

BrTarget

NameTag

Operator

Signedness

Whether to interpret an integer as signed or unsigned

Size

Type

Value

A constant value embedded in the instructions

Constants

F32

F64

I32

I64

SF32

SF64

SI32

SI64

SU32

SU64

Traits

IntoType

MicrowasmReceiver

WIP: Trait to abstract over either producing a stream of Microwasm or directly producing assembly from the Wasm. This should give a significant speedup since we don’t need to allocate any vectors or pay the cost of branches - we can just use iterators and direct function calls.

Functions

dis

Type Definitions

OperatorFromWasm

SignfulType

SignlessType

WasmLabel