Enum lightbeam::microwasm::Operator

pub enum Operator<Label> {
    Unreachable,
    Block {
        label: Label,
        params: Vec<SignlessType>,
        has_backwards_callers: bool,
        num_callers: Option<u32>,
    },
    Label(Label),
    Br {
        target: BrTarget<Label>,
    },
    BrIf {
        then: BrTargetDrop<Label>,
        else_: BrTargetDrop<Label>,
    },
    BrTable(BrTable<Label>),
    Call {
        function_index: u32,
    },
    CallIndirect {
        type_index: u32,
        table_index: u32,
    },
    Drop(RangeInclusive<u32>),
    Select,
    Pick(u32),
    Swap(u32),
    GlobalGet(u32),
    GlobalSet(u32),
    Load {
        ty: SignlessType,
        memarg: MemoryImmediate,
    },
    Load8 {
        ty: SignfulInt,
        memarg: MemoryImmediate,
    },
    Load16 {
        ty: SignfulInt,
        memarg: MemoryImmediate,
    },
    Load32 {
        sign: Signedness,
        memarg: MemoryImmediate,
    },
    Store {
        ty: SignlessType,
        memarg: MemoryImmediate,
    },
    Store8 {
        ty: Size,
        memarg: MemoryImmediate,
    },
    Store16 {
        ty: Size,
        memarg: MemoryImmediate,
    },
    Store32 {
        memarg: MemoryImmediate,
    },
    MemorySize {
        reserved: u32,
    },
    MemoryGrow {
        reserved: u32,
    },
    Const(Value),
    Eq(SignlessType),
    Ne(SignlessType),
    Eqz(Size),
    Lt(SignfulType),
    Gt(SignfulType),
    Le(SignfulType),
    Ge(SignfulType),
    Add(SignlessType),
    Sub(SignlessType),
    Mul(SignlessType),
    Clz(Size),
    Ctz(Size),
    Popcnt(Size),
    Div(SignfulType),
    Rem(SignfulInt),
    And(Size),
    Or(Size),
    Xor(Size),
    Shl(Size),
    Shr(SignfulInt),
    Rotl(Size),
    Rotr(Size),
    Abs(Size),
    Neg(Size),
    Ceil(Size),
    Floor(Size),
    Trunc(Size),
    Nearest(Size),
    Sqrt(Size),
    Min(Size),
    Max(Size),
    Copysign(Size),
    I32WrapFromI64,
    ITruncFromF {
        input_ty: Size,
        output_ty: SignfulInt,
    },
    FConvertFromI {
        input_ty: SignfulInt,
        output_ty: Size,
    },
    F32DemoteFromF64,
    F64PromoteFromF32,
    I32ReinterpretFromF32,
    I64ReinterpretFromF64,
    F32ReinterpretFromI32,
    F64ReinterpretFromI64,
    Extend {
        sign: Signedness,
    },
}

Variants

Unreachable

Explicit trap instruction

Block

Define metadata for a block - its label, its signature, whether it has backwards callers etc. It is an error to branch to a block that has yet to be defined.

Fields of Block

label: Labelparams: Vec<SignlessType>has_backwards_callers: boolnum_callers: Option<u32>

Label(Label)

Start a new block. It is an error if the previous block has not been closed by emitting a Br or BrTable.

Br

Unconditionally break to a new block. This the parameters off the stack and passes them into the new block. Any remaining elements on the stack are discarded.

Fields of Br

target: BrTarget<Label>

Returning from the function is just calling the "return" block

BrIf

Pop a value off the top of the stack, jump to the else_ label if this value is true and the then label otherwise. The then and else_ blocks must have the same parameters.

Fields of BrIf

then: BrTargetDrop<Label>

Label to jump to if the value at the top of the stack is true

else_: BrTargetDrop<Label>

Label to jump to if the value at the top of the stack is false

BrTable(BrTable<Label>)

Pop a value off the top of the stack, jump to table[value.min(table.len() - 1)]. All elements in the table must have the same parameters.

Call

Call a function

Fields of Call

function_index: u32

CallIndirect

Pop an i32 off the top of the stack, index into the table at table_index and call that function

Fields of CallIndirect

type_index: u32table_index: u32

Drop(RangeInclusive<u32>)

Pop an element off of the stack and discard it.

Select

Pop an i32 off of the stack and 2 elements off of the stack, call them A and B where A is the first element popped and B is the second. If the i32 is 0 then discard B and push A back onto the stack, otherwise discard A and push B back onto the stack.

Pick(u32)

Duplicate the element at depth depth to the top of the stack. This can be used to implement LocalGet.

Swap(u32)

Swap the top element of the stack with the element at depth depth. This can be used to implement LocalSet.

GlobalGet(u32)

GlobalSet(u32)

Load

Fields of Load

ty: SignlessTypememarg: MemoryImmediate

Load8

Fields of Load8

ty: SignfulIntmemarg: MemoryImmediate

Load16

Fields of Load16

ty: SignfulIntmemarg: MemoryImmediate

Load32

Fields of Load32

sign: Signednessmemarg: MemoryImmediate

Store

Fields of Store

ty: SignlessTypememarg: MemoryImmediate

Store8

Fields of Store8

ty: Size

ty on integers

memarg: MemoryImmediate

Store16

Fields of Store16

ty: Size

ty on integers

memarg: MemoryImmediate

Store32

Fields of Store32

memarg: MemoryImmediate

MemorySize

Fields of MemorySize

reserved: u32

MemoryGrow

Fields of MemoryGrow

reserved: u32

Const(Value)

Eq(SignlessType)

Ne(SignlessType)

Eqz(Size)

eqz on integers

Lt(SignfulType)

Gt(SignfulType)

Le(SignfulType)

Ge(SignfulType)

Add(SignlessType)

Sub(SignlessType)

Mul(SignlessType)

Clz(Size)

clz on integers

Ctz(Size)

ctz on integers

Popcnt(Size)

popcnt on integers

Div(SignfulType)

Rem(SignfulInt)

And(Size)

Or(Size)

Xor(Size)

Shl(Size)

Shr(SignfulInt)

Rotl(Size)

Rotr(Size)

Abs(Size)

Neg(Size)

Ceil(Size)

Floor(Size)

Trunc(Size)

Nearest(Size)

Sqrt(Size)

Min(Size)

Max(Size)

Copysign(Size)

I32WrapFromI64

ITruncFromF

Fields of ITruncFromF

input_ty: Sizeoutput_ty: SignfulInt

FConvertFromI

Fields of FConvertFromI

input_ty: SignfulIntoutput_ty: Size

F32DemoteFromF64

F64PromoteFromF32

I32ReinterpretFromF32

I64ReinterpretFromF64

F32ReinterpretFromI32

F64ReinterpretFromI64

Extend

Fields of Extend

sign: Signedness

Implementations

impl<L> Operator<L>

pub fn is_label(&self) -> bool

pub fn is_block(&self) -> bool

pub fn end(params: Vec<SignlessType>, label: L) -> Self

pub fn block(params: Vec<SignlessType>, label: L) -> Self

pub fn loop_(params: Vec<SignlessType>, label: L) -> Self